Updated CMake script and codebase to fix compilation on MSVC (/Clang-cl)
This commit updates CMake scripts and configs to fix compilation on
Windows with either MSVC or Clang-cl (MSVC based Clang).

+++ src/base/trace_event/trace_event_register.h
@@ -99,7 +99,7 @@ class EventRegister {
 
   uint64_t GetTs() {
     static const auto origin = base::Time::Now();
-    return (base::Time::Now() - origin).InMicroseconds();
+    return static_cast<uint64_t>((base::Time::Now() - origin).InMicroseconds());
   }
 
   uint64_t GetPid() { return TracePlatform::GetPid(); }
